Organize flat lib helpers into domain folders.
Group inventory, supabase, pdf, utils, and auth helpers under src/lib subfolders and update imports accordingly.
