+++ src/components/page-builder/record-detail-modal/useRecordDetailModal.ts
@@ -6,8 +6,8 @@ import { apiClient, membershipService } from '@/lib/api';
 import { crmRecordsApi } from '@/lib/api/services/crmRecords';
 import { useAuth } from '@/hooks/useAuth';
 import { ALLOWED_STATUSES } from '@/constants/inventory';
-import { formatCurrencyInputLive, formatPriceForInput, PRICE_FIELD_KEYS } from '@/lib/currencyFormat';
-import { getInventoryWorkflowButtons } from '@/lib/inventoryWorkflow';
+import { formatCurrencyInputLive, formatPriceForInput, PRICE_FIELD_KEYS } from '@/lib/utils/currencyFormat';
+import { getInventoryWorkflowButtons } from '@/lib/inventory/workflow';
 import type { StatusActionWithWarningConfig } from '@/components/config_components/StatusActionWarningModal';
 import type { RequestHistoryEntry } from '@/components/page-builder/RequestHistoryPanel';
 import type { RecordDetailModalProps } from './types';
