+++ src/features/billing/pages/BillingPage.tsx
@@ -13,7 +13,7 @@ import {
   downloadPdfBytes,
   PDF_LETTERHEAD_BOTTOM_MARGIN_MM,
   PDF_LETTERHEAD_TOP_MARGIN_MM,
-} from '@/lib/pdfLetterhead';
+} from '@/lib/pdf/letterhead';
 import { toast } from 'sonner';
 import {
   useBillingReport,
