+++ src/lib/inventory/priority.ts
