+++ src/lib/inventory/statusStyles.ts
