+++ src/lib/supabase/index.ts
@@ -1,0 +1,1 @@
+export { supabase, getSupabaseRestConfig } from './client';
