+++ src/lib/utils/safeProfileImageUrl.ts
